+++ packages/next/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @vercel/next
 
+## 4.6.0
+
+### Minor Changes
+
+- Support process tracing ([#12894](https://github.com/vercel/vercel/pull/12894))
+
+### Patch Changes
+
+- Fixed bug where rewrites would be applied twice in some cases ([#13040](https://github.com/vercel/vercel/pull/13040))
+
+- pin fixtures and set up auto updates ([#12979](https://github.com/vercel/vercel/pull/12979))
+
+## 4.5.2
+
+### Patch Changes
+
+- Added new prefetch segments feature for Next.js ([#12897](https://github.com/vercel/vercel/pull/12897))
+
 ## 4.5.1
 
 ### Patch Changes